Refactor functions directory and make it a Python package
- Added __init__.py to transform the functions directory into a Python package
- Renamed utility files for better descriptiveness and consistency
