+++ functions/noise_level_utils.py
